app/requirements.txt
asgiref==3.8.1
Django
sqlparse==0.5.0
typing_extensions==4.12.2
djangorestframework
psycopg2-binary
django-cors-headers
